Move class to portfolio package
